Prevent deletion of used parameters
Prior to this change, a parameter which was still in
use could be deleted. With that the content pack could not
be parsed anymore, since the information of the original
is lost.

This change will check if a parameter is still in use. If the
parameter is in use the delete button will be disabled.

Also rename apply to use / or edit to make it clearer to the
user what is going on.