Display to the user if a Entity comes from Server
Prior to this change, a user could not distinquish
if a entity came from server or from the original
content pack. This information becomes necessary
since a parameter can't be removed from a entity
not from server.

This change adds an icon to the EntitiesList showing
if a entity is from the server or the original
content pack.