Prevent clearing  of parameters which came with content pack
Prior to this change, a parameter could (nearly) be cleareds
from the content pack but since there is no value to be set instaed
the original parameter was set again.

This change will remove the "Clear" button from the gui if
the parameter is coming from the original content pack.