Prevent clearing of parameters origin from Content Pack (#5488)
* Prevent deletion of used parameters

Prior to this change, a parameter which was still in
use could be deleted. With that the content pack could not
be parsed anymore, since the information of the original
is lost.

This change will check if a parameter is still in use. If the
parameter is in use the delete button will be disabled.

Also rename apply to use / or edit to make it clearer to the
user what is going on.

* Display to the user if a Entity comes from Server

Prior to this change, a user could not distinquish
if a entity came from server or from the original
content pack. This information becomes necessary
since a parameter can't be removed from a entity
not from server.

This change adds an icon to the EntitiesList showing
if a entity is from the server or the original
content pack.

* Prevent clearing  of parameters which came with content pack

Prior to this change, a parameter could (nearly) be cleareds
from the content pack but since there is no value to be set instaed
the original parameter was set again.

This change will remove the "Clear" button from the gui if
the parameter is coming from the original content pack.

* Display actual use of a parameter

Prior to this change, the use flag of a parameter
was only displayed if it came from the original content pack.

This change uses the appliedParameter list, which also contains
freshly add ueses.

Fixes #5426