Disabling double click in plotly graphs. (Graylog2/graylog-plugin-enterprise#637)
This change disables resetting a plotly graph using a double click. Due
to our own handling of zoom events, this lead to a zoom to a minimal
time range, effectively nulling the graph and the view.

The change adds a check in the relayout handler to catch double click
events, which are being passed as empty object ('{}'). In this case no
zooming is performed.

Additionally, there is a hint from plotly shown on first zoom
("Double-click to zoom back out") which is disabled via CSS, as it is
not possible to do this otherwise.

Fixes Graylog2/graylog-plugin-enterprise#473.