Stub CSS modules for flow. (Graylog2/graylog-plugin-enterprise#652)
This PR is adding stubs and typings replacing CSS imports for flow. It
is differentiating between generic css imports (every import ending in
`.css`) and useable styles (imports starting with `!style/useable`
ending with `.css`). The latter contain typings for the `use`/`unuse`
functions as well, so errors where those are used but not properly
imported are detected too.