Fix handling of packets that contain only templates
The existing aggregation code only adds templates to the aggregated
packet, that are referenced in flow records of the current packet.
Some implementations (Fortinet) send template-only packets, leading
to an aggregated packet that does not include any templates.

Change the aggregation loop to also add the templates that
the buffered packets we are gonna return reference.