Replace handling of unhandled VARINT types with :skip
When parsing records with `application_id` or `acl_id_asa`
the code would run into the default path, because the
VARINT type is not handled. This would not advance the read index
and would parse garbage for every field that follows.

Replace the VARINT type with SKIP.
Neither of the fields that use the VARINT type, are actually
varint encoded.
Simply skip over the fields we cannot parse.

We might add application-id parsing in another change.

Fixes #4925