Fix protect still in used entites from uninstall
Prior to this change, a grok pattern could be uninstalled
even though it was still in use by another installation.

This change fixes the search to detect a usage of a grok pattern.
This was wrongly done by searching with the content_pack_entity_id
which is variable for every installaion. The native entity id
must be used (which is the id of the installed object). This
will be stored when either creating the new entity or
when finding the existing one.

Fixes: #5399