Fix endless loop after deleting a widget from a dashboard (#5500)
* Fix endless loop on widget deletion

Check for null DOM node before wrapping it with jQuery, as that will
always return an object, even if the given argument is `null` or
`undefined`.

Fixes #5498

* Do not recalculate widget size after delete

Once the widget is marked as deleted, it won't be rendered any more and
its `ref` will be gone. At that point, we don't need to recalculate the
widget size.