Fix widget creation modal with custom configuration (#5510)
* Use JSX to create custom configuration element.

* Fix linter errors

* Fix widget creation modal with custom config

Due to performance improvements in 3.0, components rendered inside a modal
are only mounted when they will be visible. That broke the way that
`WidgetCreationModal` extends its initial state with custom configuration
set in plugins, as it required components to be mounted when the modal was
opened and this is no longer the case. It is also not easy to know when
custom components are mounted to get their initial configuration, so
we need to make this code more robust to make this work reliable.

This commit changes how custom create configuration can set their
initial configuration. There are two ways of achieving this:

- Set the `initialConfiguration` static property
  This is the preferred way of setting initial configuration, and it is
  used by most of our components extending the widget creation modal,
  as the configurations are static and known when writing the code.

- Set derived configuration with `setInitialConfiguration`
  This is meant to allow components extending the widget creation modal
  to set initial configuration that is derived from props or state. As it
  sets the initial configuration, this should be called in the
  `constructor` or `componentDidMount` of the component.

It is also possible to combine both methods to set the initial
configuration. In case of duplicated configuration keys, the ones passed
through `setInitialConfiguration` will override the ones set in
`initialConfiguration`.

Fixes #5499

(cherry picked from commit 258d527ffb668e9991bedcbf4f70b997c04ba811)