Protect entities which exisited prior content packs
Prior to this change, a entity (like a grok pattern) might
have been created by a user and used (like in a pipeline rule).
Now the user installs a content pack containing exactly
the same grokpattern (name and value). If the user
uninstalled the content pack later on, he would delete
the entity which was on the system without the content pack.

This change will flag an entity if it was on the system
prior installation of a content pack and will not
deltet the entity on uninstall but rather only delete
the content pack meta data.