Add action to remove field from current table. (Graylog2/graylog-plugin-enterprise#672)
* Add message field to widget the action was triggered from.

Before this change, the "Add to table" action triggered for a message
field added this field to the static message table. After this change,
it adds it to the static message table if it was triggered from it or
from the sidebar, but if it was triggered from inside a message table
widget, it is added to this one.

* Add action to remove field from current table.

This change adds a field action that removes that field from the current
message table. If it is called from the field sidebar, the field is
removed from the static message list.

If the field is not present in the current widget, the action is
disabled.

Fixes Graylog2/graylog-plugin-enterprise#472.