Removing own AMQP reconnecting in favor of the library's.
Before this change, the `AMQPConsumer`, triggered by a connection
shutdown callback, tried to reconnect to the broker, supposedly because
the library did not support auto recovery at the time of implementation.

Current version of the rabbit AMQP client library do support it and in
the version used it is enabled per default. This lead to creating
another connection needlessly, wasting resources and leaving one of the
connections in a state that it is operating on a closed channel,
therefore throwing an exception and NACKing the first incoming
exception, ceasing operation afterwards.

This change is doing these things to improve the situation:

  * Enable auto recovery explicitly in case the default changes in the
future
  * Removing custom reconnection code, leaving only a short logging call

Fixes #4474.