Invert "add beats prefix" option
This option was added to keep the Beats backwards compatible.
Later it was decided to fork the Beats input altogether,
so we don't have to worry about breaking compatibility.

Therefore invert and enable the Beats prefix option by default
on the new input.
Having the prefix avoids field collisions.
One example is that filebeat would fill the `source` field to the
filename, not to the hostname, which is more graylog like.