Fix annotations from bernd
Prior to this change, the found_on_system was nullable which is
not necessary. The function countInstallationOfEntityByIdAndFoundOnSystem
was loading all Installations and filtered in memeory.

This change removes the nullable annotaion from found_on_system.
Uses DBQuery to query the count of installations with found_on_system.
Also a comment was improved!