Return 404 if invalid object id is specified. (Graylog2/graylog-plugin-enterprise#687)
This change fixes an issue where specifying an invalid id (as in not
being a valid ObjectId) would lead to an infinite loading screen due to
a 500 returned from the backend. With this change, the backend returns a
proper 404 which is handled in the frontend by showing the "Not found"
page.