Use default_value of parameter for entity title
Prior to this change, a entity with a parameter as title
did display the parameter name as a title.

This change tries to get the default value of the parameter
to dispaly as title. If no default value is given
the fallback is still the name of the parameter.

Fixes #5332