Use default_value of parameter for entity title (#5552)
* Use default_value of parameter for entity title

Prior to this change, a entity with a parameter as title
did display the parameter name as a title.

This change tries to get the default value of the parameter
to dispaly as title. If no default value is given
the fallback is still the name of the parameter.

Fixes #5332

* Fix Creation of content pack

Prior to this change, the creation of a content pack
was broken, since no fetched entities could be created.

This change fixes the call to pass the parameters to the
entities.

* Fix display of default value when creating a content pack

Prior to this change, the entity list could not
dispaly the default value of a parameter when
showing the preview of a content pack.

This change adds the needed parameters to the content packs
when switching to the content pack preview site.

* Fix handling of null values in entity data

Prior to this change, the dataValueIsParameter could not
handle if the data is a null value (which is possible for example for
reports).

This change adds a guard and a test to ensure the gui is more
robust.