Use rule id for reference instead of name in content pack
Prior to this change, a pipeline would search for a rule
with the name of the rule to resolve the entity. Since
the rule reference is the name of the rule.

This change will search for the actual rule by name and
uses its id as a entity reference.

Fixes #5572