Do not duplicate message in Beats input (#5582)
If the new Beats input was configured to add the
beats type as a prefix to every field (the default),
it would duplicate the entire message field to
<beats-type>_message.

Fix this and additionally, keep the `gl2_source_collector` field
from being prefixed. This is needed for the "show messages" button on
the sidecar view.

(cherry picked from commit d64f016373ae4209575c3396776408c6f33d699f)