Reset dirty flag when existing view is updated. (Graylog2/graylog-plugin-enterprise#744)
* Reset dirty flag when existing view is updated.

Usually the dirty flag of the view is set to `false` when the view is
loaded, which also happens when saving it for the first time. This
change performs this as well, when an existing view is saved. This
prevents the window/route leave handler to nag the user to save the
view even if it was recently saved.

* Adding/adapting tests.