Fixes for index set field type refresh interval value (#5595)
* Fix linter errors

* Add default value to `field_type_refresh_interval`

Use same default value as the server.

Fixes #4724

* Allow restricting units in TimeUnitInput

This allows consumers of the component to only display supported units
in the form.

* Ensure we do not modify the state directly

* Ensure we render the current state in the form

The state is initialized from the props, but changes in the form are
persisted in the state. To ensure any computed value from state is still
correctly rendered, use the state values instead of the props.

* Restrict available units to seconds and minutes

* Rely on moment.duration to get right values

On change, always convert the given value and unit to milliseconds, which
is what the server expects.

For rendering, convert the given time in milliseconds to the selected
unit.

* Check fieldTypeRefreshInterval 1 second or longer

Refreshing field type information too often may have a big performance
impact. This change ensures the refresh interval is not shorter than one
second, which seems to be a reasonable low value for it.

(cherry picked from commit 198cc699f6a14493278c9ffcf42e895e8325b9cb)