Assign new search id upon creation. (Graylog2/graylog-plugin-enterprise#749)
* Assign new search id upon creation.

Before this change, a new search id had to be assigned explicitly, which
lead to reusing the same search id even if modifying an existing search.

This change is now assigning a new search id before it is created in the
backend. This should prevent situations where the search is modified but
not reflected in the view. Especially widget mappings are sensible when
it comes to changes in the search due to mismatches of search type ids.

* Removing unnecessary change.