Fix changing index rotation and retention settings (#5622) (#5624)
The fixes in #5595 corrected the code in `IndexSetConfigurationForm` to
avoid modifying the React state directly. Creating a clone of the state
had an unexpected consequence: the `_updateConfig()` method updates
different object keys stored in the `indexSet` state key. When
consecutive calls to `_updateConfig()` are made, each of them will
create a clone of the state at the time of the call, but that state does
not represent the actual current state, since React enqueues changes to
the state.

In order to avoid that issue, we make use of the update function to set
the state, which guarantees that the current state (including any queued
modifications) are passed to the function.

Fixes #5617

(cherry picked from commit aad8a4d51503155afc3cc03d0c904f237b0c3127)