Fix slow loading UI after server startup
Avoid registering ShiroSecurityContextFilter as a DynamicFeature.
Doing so, triggers Jersey to create a ServiceLocator for all
existing (~1600) Resources. [1]
Furthermore, each created ServiceLocator will have its own Guice/HK2 Bridge,
whose JIT resolver cache is cold upon startup and needs to be filled.

Instead of registering the ShiroSecurityContextFilter dynamically, we
register it statically in the JerseyService because every Resource is filtered
through it anyway.
We need to give it a higher priority so it's executed before the authentication filter.

[1]
https://github.com/jersey/jersey/blob/2.25.1/core-server/src/main/java/org/glassfish/jersey/server/model/ResourceMethodInvoker.java#L187