Fix test depending on Graylog minor version.
The test for the `AppErrorBoundary` was using a DOM snapshot containing
verbatim HTML elements generated by the `ErrorPage` component including
the current Graylog version. This was leading to a test failure after
the minor version was bumped.

This change is replacing the snapshot with test matchers checking for
the existence of the `ErrorPage` component and the correctness of the
props passed to it.