Unregister route leave handler when unmounting `WindowLeaveMessage`. (Graylog2/graylog-plugin-enterprise#795)
Before this change, the `RouteLeaveHook` which is registered on the
router upon mounting, was not unsubscribed. This lead to the handler
being called on other pages when the route was changed as well.

This change is now retaining the returned unsubscribe function and calls
it upon component unmount.

Fixes Graylog2/graylog-plugin-enterprise#792.