Fix http input keep-alive handling
We were only responding to the first request on an http keep-alive connection
with a 202.
Any following requests would not be answered, leading to a timeout in the client.

Configuring a ChannelFuture for CLOSE_ON_FAILURE did expose the reason
for the bug:

`io.netty.handler.codec.EncoderException:
java.lang.IllegalStateException: unexpected message type:
DefaultHttpResponse, state: 1`

We ought to use `DefaultFullHttpResponse` instead of `DefaultHttpResponse`.
Also keep the `CLOSE_ON_FAILURE` to make this more robust for future bugs.

Fixes #5720