Add MessageOutput.Factory2 interface for guice
New outputs need to be able to receive the Output instance via assisted
inject. We cannot just change the existing MessageOutput.Factory
interface because that would break older output plugins.

Using a "default" method for the new "create()" method also doesn't work
because guice cannot handle that and old plugins would fail to load.

The only way seems to be the introduction of a second Factory interface
that can be used by newer outputs to get the Output instance.

This change also removes "MessageOutputFactory#get(String)" because it
doesn't seem to be used anymore and it cannot work with two different
output factories.