Prevent unnecessary updates of `MapVisualization` component. (Graylog2/graylog-plugin-enterprise#814)
Before this change, the state of the viewport was tracked in the states
of both the `WorldMapVisualization` and the `MapVisualization`
components. This lead to unrequired updates of the `MapVisualization`
and triggered cascading `onViewportChanged` calls, finally resulting in
too many component state updates.

After this change, the state is tracked in the `WorldMapVisualization`
component only and propagated through to the `MapVisualization` if
needed.

Fixes Graylog2/graylog-plugin-enterprise#812.