Do not include '*' when generating new query. (Graylog2/graylog-plugin-enterprise#826)
Before this change, any new query that was generated contained a query
string of `*` by default. This was unnecessary, as the empty string has
the same meaning for lucene. Additionally, the placeholder text of the
query input was not seen by default, only if it was emptied.

This change is removing the wildcard and using an empty string for the
query by default.