Do not try to close completion popup when executing query if not present (Graylog2/graylog-plugin-enterprise#832)
In Graylog2/graylog-plugin-enterprise#257, it was introduced to close the popup when executing a query
using the enter key. Unfortunately, it was not checked if there is
actually a popup (for example when a query was pasted), leading to an
exception and the query execution is not taking place.

This change is adding a guard, checking if there is a popup and hiding
it in this case.