Server changes for Forwarder (#5670)
* Use class-based unique metric names

Allows a subclass instance to run without throwing a "metric already exists" exceptions. This will need to be refactored later to support multiple instances. Skipping this for now to prevent CI errors in enterprise integrations repository.

* Adjust comment

* Add additional MessageOutputFactory constructor including output

This allows us to access the output ID in a backwards compatible way (does not break existing 3rd party outputs)

* Make KafkaJournal metrics unique across instances

The KafkaJournal had several hard-coded metrics directly referencing the KafkaJournal.class class name. This have been changed to allow a metricPrefix override, which can allow instances to provide unique metric names for each instance.

* Add hack for GlobalMetricName backwards compatibility

Also revert unintended formatting changes.

* Use correct oldest-segment metric name

* Refactor global oldest-segment metric name to avoid duplication

* Fix bug when checking metric name

* Revert unintended changes to KafkaJournalTest

* Teardown metrics in shutdown method

Prevents duplicate metric error message when restarting the Forwarder (via Outputs page > Edit > Save)

* Add MessageOutput.Factory2 interface for guice

New outputs need to be able to receive the Output instance via assisted
inject. We cannot just change the existing MessageOutput.Factory
interface because that would break older output plugins.

Using a "default" method for the new "create()" method also doesn't work
because guice cannot handle that and old plugins would fail to load.

The only way seems to be the introduction of a second Factory interface
that can be used by newer outputs to get the Output instance.

This change also removes "MessageOutputFactory#get(String)" because it
doesn't seem to be used anymore and it cannot work with two different
output factories.

* Add MessageOutput.Factory2 handling to OutputFacade in content-packs

* Teardown all metrics on shutdown

Several metrics were not being torn down at shutdown time. These have been cleaned up. The setup of metrics was left in two blocks (did not want to consolidate and potentially cause a situation where a metric is not created yet when used in the constructor)

* Add comment explaining the factory order in MessageOutputFactory

* Add support for disabling KafkaJournal throttling

According to a comment in graylog.conf, the original intent of the KafkaJournal.throttleThresholdPercentage field was to disable throttling updates if the value is not specified. This commit changes KafkaJournal to not send throttle state changes when this value is not specified.

* Use -1 instead of null to indicate that throttling should be disabled

* Prevent race condition between Output.stop() and doGracefulShutdown()

This solves an intermittent race condition between `ForwarderOutput.doGracefulShutdown()` and the `ForwarderOutput.stop()` method. Upon server shutdown, both methods were being called. If `stop()` executed first, then the graceful shutdown was suppressed (due to the shutdownInProgress AtomicBoolean), which resulted in an incomplete shutdown.

The shutdownInProgress AtomicBoolean and IllegalStateException catch block were also removed, as they are no longer needed.