Allowing to retrieve current state from `InputsStore`.
Before this change, the `InputsStore` did not contain a
`getInitialState` function, allowing consumers (especially `connect`) to
retrieve the current state without triggering an update.

This is required by `connect` to get its initial seed. Also it
alleviates the need to ensure that the consumer is listening to the
store _before_ a `list` action is triggered.