Fixing usage of `undefined` for inputs due to new initial store state. (Graylog2/graylog-plugin-enterprise#892)
Due to the change Graylog2/graylog2-server#5791, the `MessageList` gets
a new initial state. Instead of having `undefined` for the `inputs` prop
on first render, it is now `{ inputs: undefined }`, so it is not
replaced with the default prop and therefore using `map` on
`this.props.inputs.inputs` fails.

This change is assigning an empty array as default to the destructured
`inputs` variable to enable `map`.