Add RefreshControls to Searchbar (Graylog2/graylog-plugin-enterprise#876)
* Add RefreshControls to Searchbar

The new views were missing a way to automatically refresh the
views page.

This change, adds the refresh controls from the old search
and wires them to the new.

* Move state into store and use actions

Prior to this change, the state was kept inside of the
component. But other components needed to stop the
refresh as well.

This change moves the state into a store and calls it when
the auto refresh will start anoying the user.

* Merge RefreshActions into RefreshStore

In some cases it can be unwise to seprate the RefreshActions from the RefreshStore.
When the RefreshActions are called before the RefreshStore is loaded.

Also add RefreshActionTypes to the RefreshActions.

* Fix Flow check bugs

* Disable auto refresh when changing the page

Also, fix enable state when changing the interval
while beeing disabled.