Add support for extractor creation links (Graylog2/graylog-plugin-enterprise#871)
* Add support for extractor creation links

Prior to this change, there was now way to create extractors inline
from the extended search.

This change adds a new action to the values action. Which opens a modal
asking for the type of the extractor. After selecting a type it opens
a new window where a extractor with the example value can be created.

Fixes: Graylog2/graylog-plugin-enterprise#865

* Move SelectExtractorType into logic/valuactions