Deactivate automatically sorting by plotly (Graylog2/graylog-plugin-enterprise#887)
* Deactivate automatically sorting by plotly

Prior to this change, numerical pivots were sorted
by plotly automatically. Even though the user already
set a sorting, plotly did a sorting by itself.

This change treats the data of the xaxis as 'category'
which will not be sorted from plotly. That way the data
is represented as it comes in. (Already sorted by us).

Fixes Graylog2/graylog-plugin-enterprise#864

* Fix failing test

* Only disable plotly sort, if sort is done by us