Make `Timestamp` accept seconds from epoch.
Before this change, passing a number containing a timestamp based on
seconds after epoch resulted in a proper output, but also raised a prop
type validation error.

This change is now adding `number` to be a valid prop type for
`dateTime` and documents the possibility to use seconds after epoch for
`Timestamp`.