Move web build and cache directory into target/
The "build" and "cache" directories are currently located in the project
root of plugins. This change moves them to the "target/web" directory.

That way we don't need any custom configuration to remove these
directories on "mvn clean" and they should also not be indexed by
IntelliJ and consume resources. (most developers exclude the "build" and
"cache" directories in IntelliJ for performance reasons)