Reset dirty flag when creating view. (Graylog2/graylog-plugin-enterprise#927)
When creating a new view, the dirty flag is set to `false`, but
afterwards a new search is generated and loaded, which triggers setting
`dirty` to `true`. This lead to unnecessary prompts when leaving the
extended search page after creating a new view, but without modifying
it.

This change is now setting the dirty flag to `false` _after_ creating
and loading the search.