Fix onChange handling in `WorldMapVisualization`. (Graylog2/graylog-plugin-enterprise#939)
In Graylog2/graylog-plugin-enterprise#932, refactoring of the `WorldMapVisualization` component introduced
a regression, that lead to markers on the map not being drawn due to
triggering the `onChange` callback, which did not have a function
referenced when not in edit mode. The missing function reference was
related to a default value for a prop being removed during refactoring.

This change is now calling the `onChange` only when in editing mode,
fixing the drawing of the markers. Additionally, it is restoring the
default value for the prop.