Adding more rules to Flow strict mode. (Graylog2/graylog-plugin-enterprise#941)
When flow was introduced in this plugin, `strict` mode was used in files
where typing was introduced. Without further configuration, strict mode
does not differ from non-strict mode regarding the reported errors. Now
that flow is established, this change is adding three more rules to
strict mode:

  * `untyped-type-import`: importing types from untyped files is not
allowed
  * `unnecessary-invariant`: using `invariant` on conditions which are
inferred to be always `true` is reported
  * `sketchy-number`: using a number as condition is reported, due to it
being potentially defined but falsy (`0`)

Additionally, `memoize-one` is declared to be untyped, due to errors
because of missing flow types.