Stub simple backend, wait for networkidle0.
Before this change, the script was neither waiting for the page to
finish network requests and was not stubbing the simplest parts of the
backend api. This lead to different behavior based on timing
differences. On one machine it returns after loading the page
successfully, on other machines it was already trying to query the
backend (unsuccessfully).

This change is now:

  * Adding a simple backend api stub on a random port, responding to the
root uri and the session api
  * Waiting for networkidle0 (no network requests for 500ms)