Checking health of production build assets during compilation. (#5894)
* Adding basic script to check health of production assets.

* Improving error handling.

* Loading script assets from HTML, not explicitly.

Loading assets explicitly using `addScriptTag` was never firing the
`window.load` event.

* Load plugins from module.json passed via arguments.

* Returning proper exit code based on results.

* Adding yarn task to check production asset health.

* Adding task checking production asset health to maven run.

* Bomb out in case of bootstrapping errors.

* Downloading headless_shell and using it if on linux.

* Correcting prop name.

* Removing console.log statement.

* Stub simple backend, wait for networkidle0.

Before this change, the script was neither waiting for the page to
finish network requests and was not stubbing the simplest parts of the
backend api. This lead to different behavior based on timing
differences. On one machine it returns after loading the page
successfully, on other machines it was already trying to query the
backend (unsuccessfully).

This change is now:

  * Adding a simple backend api stub on a random port, responding to the
root uri and the session api
  * Waiting for networkidle0 (no network requests for 500ms)

* Fixing linter hints.