Replace connecting component to `InputsStore` with single action call.
Last commits were fixing the issue only partly. It turns out, that the
`InputsStore` relies on an internal state (calling `get` with an input
id) to return the correct input in the `input` key of its state.
Therefore the extractors page was returning the extractors of the last
opened input instead of the current one.

This change is replacing the `connect` call with a single call to the
`get` action and sets the current input in the callback of the promise.