Do not initialize `inputs` variable to prevent merging duplicate key in. (#5954)
* Do not initialize `inputs` variable to prevent merging duplicate key.

This change is fixing an issue that is related to the introduction of
its supplied state even during first initialization, leading to an
attempt to merge a duplicate key when mounting the `ExtractorsPage`
component. Removing the `inputs` key in the initial state of that
component is fixing it.

Fixes #5948.

* Fixing linter hint by simplifying node change handler.

* Replace connecting component to `InputsStore` with single action call.

Last commits were fixing the issue only partly. It turns out, that the
`InputsStore` relies on an internal state (calling `get` with an input
id) to return the correct input in the `input` key of its state.
Therefore the extractors page was returning the extractors of the last
opened input instead of the current one.

This change is replacing the `connect` call with a single call to the
`get` action and sets the current input in the callback of the promise.