Allow specifying view for `PaginatedService`.
Using the `PaginatedService` with a DTO that utilizes Jackson's JSON
views to customize (de-)serialization e.g. for persistence vs. REST does
not work currently, as a view needs to be passed to MongoJack in order
to enable views processing.

This change is adding a second constructor to `PaginatedService`
allowing consumers to specify a view class as additional parameter.