Set default value for email sender to empty value
The previous default (@example.org) is problematic for use cases where
the email server rejects sender domains which are not validated / owned
by the graylog server owners (e.g. amazon ses).

In our use case users often forgot to empty the value and left the
default in which resulted in notifications which where not sent out.

With this change the default is empty and if I understand the code
correctly the result would be that the default which is set in the
server configuration should be used.