Avoid collisions when merging `InputsStore` state into components state.
Before this PR, mounting one of the affected components lead to a
collision when merging the state of the `InputsStore`, as returned by
its `getInitialState` function into the state of the components, because
the `input` key was already defined. This PR is removing the initial
assignment of this key and fixes linter hints.

Fixes #5997.