Add support for nesed maps in list for Content Packs
Prior this change we missed support of maps in list of
a ReferenceMap. This is needed to support configurations
(like DashboardWidget) which can contain a list of maps.
(e.g series when two charts are merged into one)

This change will call toReferenceMap for every list item
which is a map.

Fixes #5742