Adding helper to ensure a given value exists only once.
During early development of Graylog web plugins it was noticed that when
a store was imported from a plugin, it was duplicated. Importing a file
from a different scope did not reuse the previous import but imported
the file again, and any action performed at import time is performed
again. This was detrimental, as stores, which were meant to centralize
data and the actions performed on them, were duplicated and actions
performed on them did only happen on at most one of the stores.

The solution was to ensure that stores/actions are singletons.

For this, the `(Action|Combined|Store)Providers` were introduced, which
ensure that importing (or in this case `require`ing) a dependency was
always performed in the same scope. The downside of this approach is
that imports were turned into function calls and, more important:
*the consumer of a dependency was held responsible to ensure that a
dependency was always aquired through this mechanism*.

This had several negative consequences:

  * Navigating in an IDE to an implementation of a store/action broke
  * Potential type information for a store/actions is lost
  * Writing a test for anything that requires a store/actions requires
mocking the provider classes, making it harder and adding lots of
boilerplate
  * All stores/actions must be registered in a central instance
  * More boilerplate was required to make this work for code splitting
(ensuring that stores/actions are in a separate bundle, being loaded
before everything else)

To address these shortcomings, the `singleton` helpers were introduced
during the open sourcing of the views feature. These allow ensuring
that a given value is initialized only once, being very useful for
exports. Any other time this value is imported, the already initialized
value will be used instead. This has the benefits of:

  * Delegating responsibility for being singleton to the exporter
  * Retaining the IDE's capability for navigation
  * Easier mocking for dependencies
  * Dependencies can be imported like any other dependency, even from
      plugins

The usage requires two elements: a `key` that must be unique for the
given value, and a `supplier function` that is evaluated lazily in case
no value for the given `key` exists. `singletonActions` & `singletonStore`
are shortcurts which are suffixing the given `key` with `Actions`/`Store`.

For now, these helpers are used only in the scope of the views, being
expanded and replacing the `*Provider` mechanism when this has proven to be
fruitful.