Exporting actions/stores of views as singletons.
This change exports all actions/stores of the views as singletons, to
make them usable from plugins as well. It uses the newly introduced
`singletonStore`/`singletonActions` helpers, as described in 45c89c5d.

For consistency and to avoid conflicts, all values are registered with a
unique key that is prefixed with `views.`.