Handle missing field type of selected field in MessageTableEntry.
Before this change, using `Immutable.List#find` to lookup a field type
used the wrong parameter count. Instead of (what the intention was)
passing a default value for cases where the field type is not present,
it was passing the `context`[0] instead. This is fixed and a test is
added.

Fixes #1090.

[0]: https://immutable-js.github.io/immutable-js/docs/#/List/find