Updating retrofit from 2.4.0 to 2.6.0.
This change is updating Retrofit from 2.4.0 to 2.6.0. The current
version is fixing a XXE vulnerability, described here[0]. It most
probably has no impact on the product, as we do not use XML in the
places where Retrofit is used, but it is unclear if e.g. a MitM could
force the Retrofit client to interpret XML and therefore trigger the
vulnerability.

The changelog for 2.6.0 is found here[1].

[0]: https://nvd.nist.gov/vuln/detail/CVE-2018-1000844
[1]: https://github.com/square/retrofit/blob/parent-2.6.0/CHANGELOG.md