Updating OkHttp from 3.11.0 to 3.14.2.
This PR is updating OkHttp to 3.14.2.

Any OkHttp version up to 3.12.0 include a vulnerability that allows an
attacker doing MitM to bypass certificate pinning. It does require full
control of the Graylog server before, as it requires the attacker to
control the startup/process space of the Graylog process though.