Fix styling differences for react select.
During the migration from `react-select` v1 to v2 some styling
differences were eliminated, but a lot were not. E.g.:

  * The dropdown caret looking different and taking up more space
  * Now having a separator between the dropdown indicator and the rest
  * Custom css fixes in `Select.css` were not taking effect anymore, due
to the different classnames and generally a different way to customize
styling, leading to:
    * z-index is not increased anymore
    * The `size` property for the component is not considered

This PR aims to remove these differences and make the component look and
behave the same as before.

Fixes #6036.
Fixes #6040.